mem: Add a DRAM cache model
This patch adds all of the code needed for a DRAM cache model. This
includes the code for multiple different DRAM cache policies including
victim dirty, victim clean, and victim adaptive, as well as a baseline
which closely follows the Intel Knight's Landing DRAM cache design.

This patch also modifies the crossbar and the packet so that the victim
policies are more efficient. This is meant to model changes in the caches,
but it was simpler to implement changes in the crossbar.

Change-Id: Ic06b10d90baecd60cd417b79208dd9a9b51a7d3e