mem, cache: Add new ReadOwnerReq memory command
ReadOwnerReq is a cache-to-cache message which requests the cache block
in an owned state (i.e., other caches can have shared copies, but they
cannot be dirty).

This patch implements the ReadOwnerReq in the cache. It depends on the
DIRTY_DATA_FROM_CACHE bit in the packet flags added by another patch
in the DRAM cache implementation.

Change-Id: Ie96a859a41ab96d6d125b81119c5339a592a1eee
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>