dramcache: Block until invalidate is complete
This solves a bug in the knl-like cache. We want to keep the line blocked
until the invalidate is complete. To do this I had to switch the
invalidates to track on the line number, not the packet address and also
add a new structure to remember what we need to invalidate.

Change-Id: Ie41ddadaaa2ca6538c2f88999424be0fe5d900b9
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>