mem: Allow abstract memories to "mirror" others.
If a memory is "mirroring" it means that it is a different implemenation
of the same (guest) physical memory. This can be used to create multiple
totally different system configurations within a single gem5 process.

Memory mirroring is implemented by syncing the mirroring memories with
the original memories and using the same backing physical memory for
both.

This is currently "alpha" level. There are a number of potential use
cases that haven't been tested. A few are listed below.
 - Checkpointing/restoring

Change-Id: Ie2208a747721cc5961ffabef1fcae43cac40f578