mem: Convert MemObjects to use connected()
This moves the sendRangeChange logic from init() into connect().

Some objects have dropped panics for when ports are unconnected since it
is now impossible for connected() to be called when the port is not
connected. This panic is left in the code for MemObjects with multiple
ports.

There are a few objects that still call sendRangeChange in init(). These
will need to migrated if they ever support hot plugging. Unchanged
objects will continue to work correctly.

Change-Id: I42d1e1d42b9a09ab13ec42476e586957ae61b056
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>