cpu: Convert CPU switching to use hotplug
This converts the CPUs to use the hotplug API instead of switchCpus.

Three tests are added to test switching timing CPUs, atomic->timing, and
atomic->simple->O3->simple->atomic with caches.

One design choice to consider: Think about making SimObject unplugged a
PyBindProperty. Currently, the CPUs mirror the unplugged property in both
Python and C++. This isn't ideal.

One thing that makes this change much simpler is the assumption that both CPU
models (the one being unplugged and the one being plugged in) want to use the
same model for the TLB and interrupts. Without this assumption, figuring out
how to hook up all of the ports and which objects need to take over becomes
very complicated. Currently, there are no (known public) cases where someone
wants to fast-forward or warmup with a different interrupt or MMU model than
what is used for detailed simulation. If this changes, we can revisit this
design choice at that time. It should be straightforward to extend this API
to the MMUs.

Required changes:
 - Convert all uses of MemoryMode enumeration to use the C++ instance, not
   strings to reduce the number of places the mapping from string->enum is
   required.

Change-Id: I03c30e429570338d1499bc37e8d8bfa290afc8aa
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>