python: Overwriting SimObjects should clear child
There are times where you want to overwrite a parameter that is a
SimObject. For instance, there may be a default for the SimObject
parameter, but you overwrite this at runtime. In some cases, the
default that was overridden was still instantiated (ARM stage 2 MMU).
This caused a fatal condition.

Now, whenever a parameter is overwritten we clear the old child before
overwriting.

Change-Id: I9ef39894fb884490e381829a25c039291b3adb38
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>