cpu: Remove the TLB takeOverFrom function
This function was a misnomer, no TLB model *actually* copied any state from
the old TLB to the new TLB. This led to at least one researcher (me)
incorrectly using switch CPUs and assuming the TLB was warm.

The function getMasterPort() is still available on TLB models. To switch to
a different TLB model when switching CPUs we can expose this to the
Python code for disconnecting and connecting ports. However, for now, we
are keeping the assumption that all CPU models will use the *same* TLB
instances.

Change-Id: I79345c78dfe82cb235634efbd9fb4083496d27fc
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>