Work in progress of replacing SimpleThread
I want to replace SimpleThread because it combines the thread context
interface that is required for things outside of the CPU to interact with
it with the hardware context state for the thread.

However, as I did this, I realized I was just re-implementing the same thing
as thread context. I think this needs to be considered more.

Change-Id: I61950a7dfd524a03991bd3ded3f643e8822a83f7
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>