flexcpu: Add resource type and refactor execution
This adds a resource type to the CPU and refactors the exeuction logic
to use the new type. This patch also significantly simplifies the logic
for requesting and completing executions. This patch uses the resources
for instructions translations, fetches, executions, data translations,
and memory requests. It uses lambdas and events probably too much, and
it should be refactored some. The framework is mostly there for limiting
bandwidth, but that hasn't been implemented, yet.

Change-Id: Id692b99aca87a16670695f09175018492f5333a1
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>