flexcpu: Reorganize front end
Now, fetching instructions, etc isn't as tied to dynamic instructions.
Also, all of the decoding now happens in one function, advanceInst.

This exposed a couple of bugs with when instructions updated the BP.

There's still a couple of things that need to be done. Because instructions
wait to update the BP until commit performance is really bad now (both
simulator and simulated performance).

Change-Id: I9aa53cc8e6377efe23c445bffcb608ed12aae57c
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>