flexcpu: Modify BP to not fetch if it's wrong
Instead of trying to fetch more instructions when we *know* the branch
predictor is wrong, simply stop the fetch unit. This will improve the
performance of the simulator (from 408s to 103s for Bubblesort). This
slightly changes the simulated system's performance, but by much less
that 1% for Bubblesort, and it's unclear if this change is anything
except random noise.

Change-Id: I370e6d38b96a2a6da79d0396444f2629892dc168
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>