cpu,arch-86: Add IsFaulting flag to StaticInst
The IsFaulting flag is set for micro-ops that are the only micro-op in
the macro-op and that fault. This lets us know at decode time if we
should keep fetching or stop. This is mostly a simulator performance
optimization, but is also useful when the number of fetches may be
unconstrained to prevent infinite loops in the simulator.

Change-Id: If7e77caeee986ed548b946667002c3af01fb8ade
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>