arch-x86: [hack!!] Fix bug in O3
When compiled with optimizations, there's cases when the O3 is trying to
decode illegal instructions (e.g., when the front end is running far
ahead) that causes a segfault in the decoder. I believe there's a case
where we say M5_UNREACHABLE when this is *not* true. If we instead emit
some reasonable code the segfault goes away.

Change-Id: I58ffea5b5eba57a6a08dd158cbff3ed8b5f5b17f