cpu-flex: Added alternative constraint for instruction window.
Common trace-based analysis tools have a configurable instruction window
within which to analyze programs. This emulates that behavior by
creating two separate constraints, (the existing) one which behaves more
like specific microarchitectures with ROBs that handle microcode (like
the ROB in O3), and a new constraint which constrains by the ISA
instructions encountered, ignoring the impact of microcode. This feature
is intended for use alongside the zero-time-microops feature, for
simulation of systems minimizing the impact of microcode, regardless of
the state of ISA implementations within gem5.

cpu-flex: Added alternative constraint to limit insts instead of ops.

Signed-off-by: Bradley Wang <radwang@ucdavis.edu>

Added some documentation, minor changes for clarity

Change-Id: Id44c369735e002e37bdb03bcdb8db3f475940dcf
Signed-off-by: Bradley Wang <radwang@ucdavis.edu>