changed the getWritableVecReg(), changed sendToMemory() callback to onCompleteExecution()
getWritableVecReg() does not use prior values, so it's unnecessary to trace back the values of old regs.
sendToMemory() has a callback to onCompleteExecution() when the load/store instruction came back from
memory. Store instructions that are not store conditional do not write back to register, so we can keep
them around until they are complete, because even if they are squashed, no younger instruction will
access that store's register. Hence the younger instructions won't access squashed instructions' results.
However, for load instructions and store conditionals, we must set them to nullptr as soon as they are
squashed, since younger instruction might try to grab the old regs from those instructions. This explains
the choice of weak_ptr for store conditionals and load instructions, and shared_ptr for non-conditional
stores.

Change-Id: I75f3dba5a30916e1034baf586be7b75d93005ab5
Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>