cpu-flex: Implement support for predicated instruction
This change determines logics for handling predicated
instuctions. We assume that the predicate will be determined
in execution stage, and older instructions that affect the
predicate will be executed before this instruction advances
to the execution stage. We handle the instructions that are
predicated to false as follows,
- If the predicated instruction is not a memory instruction, it
will go through the pipeline as a normal instruction, except
that it won't make any changes to the context. So, the value of
registers from this instruction should not be forwarded to other
instructions as the results are invalid. This will be handled in
later commits.
- If it is a memory instruction, it will go through the pipeline
as if it is a normal memory instruction, except that it won't go
to memory. Hence, effective memory address callbacks will be
called. Those callbacks must be able to handle the case where
this instruction is predicated to false. Nothing from this
instruction should be forwarded, too.
- Any type of dependency between two instructions won't be added
if the older instruction is predicated to false.

Change-Id: I89146d7f8889de441cd30033645129dd079ce33b
Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>