cpu-flex: Predicated instructions support for register forwarder
Previously, the register forwarder gets old register values from the
most recent producer of the register. However, with the recent
addition of predicated instructions, the most recent producer does
not contain the valid result when the instruction is predicated to
false and the prodcuer instruction does not go through the
execution stage. This commit changes the register forwarder such
that, the forwarded register values are either from the most recent
producer that is predicated to true, or from the backing context if
the producer instruction is committed.

Change-Id: I223c495f44bc149ea243c09d39ceb6ab00d3943a
Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>