cpu-flex: Implement getWritableVecRegOperand()
getWritableVecRegOperand() is implemented as other register set
functions since the prior values of the register are not used, and
this function is expected to return the reference to the vector
register for further direct updates by the instruction.
setAsVecReg() is added to GenericReg to set a register as a vector
register without knowing how to initialize the register. This
function should only be used when the prior values of the register
are not used.

Change-Id: I1f6ca3f1dd480161d81daadf32786473ad948ad2
Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>