cpu-flex: Add support for store conditionals
Previously, the locking operation is not called for all stores, and
stores are committed as soon as memory access begins. Different from
a normal store, store conditional performs the locking operation
before going to memory, and only goes to memory if it has the lock.
Store conditional also updates the destination register. This commit
accommodates the difference by calling the locking operation for
store conditionals, and by not committing them until memory access is
complete.

Change-Id: Ic83623f5c689cc4f03bfb5a5a7642effa9b9e452
Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>