base,sim,python: Add core python API
Add documenation to the _m5.core library and add tags to all of the
functions used in this Python API.

When executing help(_m5.core) in Python, there is now a helpful message
printed.

This and following changes are setting up the Python interface for
larger updates. The first step is to improve the documentation.

Issue-on: https://gem5.atlassian.net/browse/GEM5-647
Change-Id: I0f290eb2b689020e53cf4393ec62a9ba1725ddc9
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>