mem: Updating MemScheduler to support ruby.
The initial assumption when implementing MemScheduler was each
processor has it's dedicated port when connecting to the MemSched.
Now by using ruby all the processors are connected to the
MemScheduler through the directory connection. This patch adds a
translation layer that creates virtual cpuPortId for each
dedicated queue in MemScheduler. This translation will help
distribute incoming requests to queues by their requestorId. Also,
now there is need for a retranslation from the virtual cpuPortId
to the actual portId of the requesting port (in case there are
more than one directories connected to MemScheduler, or in more
general terms, there are more than one requesting entities
connected to the MemScheduler).

Change-Id: Ic11caec86fbacee33180732b1a5506e8cf221eee