misc: Changing mesi two level to work with test board.
This commit changes the mesi two level code for it to work with
the test board. However, it breaks the code that already worked
with full system x86. There needs to be a new board-cache interface
to be able to support boards with various ISAs. Also, the abstract
board constructor now takes cache_line_size as an input. It is defaulted
to 64 to not break already existing configurations.

Change-Id: I9bff7187133aba268c3b23e53a8aa80b87ce5a40