Adding traffic generators to component library. (#222)
* misc: Adding scripts for abstract_generator_core and simple_generator_core.

This patch adds scripts that define an abstract core for traffic generators
and a simple core that is specific to PyTrafficGen in gem5. The class
AbstractGeneratorCore has an extended interface for set_traffic and
start_traffic functions.

* misc: Updated scripts for abstract generator and abstract generator core.

* misc: Updating scripts for traffic generators

* misc: Creating first test for generator scripts.

Change-Id: Ifb1ce75a70eb9cb79d97a506dfafc1e19302ae2d

* misc: Updated scripts for generator.

Slight change in the hierarchy of generators. Still needs to
provide a better interface to the user. It works though.

Change-Id: Ib42bff2b6d9dee347c7f2692cda1a3ed26a245fe

* misc: Getting rid of helper scripts in generator implementation.

Change-Id: Ib3f3fc0f0bcd7e43b87f0ffc87b33420c31dde75

* misc: Added random generator and updated the interface.

Change-Id: If0e0d19e0e80d27bc8524389a7cc1a8f9264785d

* misc: Adding complex generator. Fixing function signatures.

Change-Id: I113ad6bae36e8699b1323ce8420ef4736376f6f0

* misc: Creating single channel configurations for gem5 and ds3.

Change-Id: Ib169d813b50ca2686424c336f1235bb78e4f68da

* misc: Adding newlines.

Change-Id: I809f7b27a29cce08cfa3268299159c89a93f18e2

* misc: Adding GUPSGen.

Change-Id: Ic34a8684b2ed513733856ca827240740a87f1312

* misc: Updating boards, adding gups generator, ...

Change-Id: I82e9babdd624c0f767126627e3cc5ac46cb57b2d

* misc: Changing mesi two level to work with test board.

This commit changes the mesi two level code for it to work with
the test board. However, it breaks the code that already worked
with full system x86. There needs to be a new board-cache interface
to be able to support boards with various ISAs. Also, the abstract
board constructor now takes cache_line_size as an input. It is defaulted
to 64 to not break already existing configurations.

Change-Id: I9bff7187133aba268c3b23e53a8aa80b87ce5a40

* misc: Removing GUPSGen, fixing styling

This commit removes py_generator_core source and brings linear_generator_core
and random_generator_core one level up the hierarchy. It also fixes some of
the styling in the code.

Change-Id: I78d03dd9d3e0067f9f61f0ed6f3ed2f0ae313601

* misc: Updating scripts for DRAMSim3.

Change-Id: I3a09a6272450277af98b77a1ff2e6e35c4d48659

* misc: Adding typing to different classes.

Change-Id: Ie654838d181da0533238a0a078d3a2ca5dd4d82c

* misc: Fixing bugs and typos.

Change-Id: I5b5d4c963ba986d74ceaba14f3730104d69e074d

* misc: Adding documentation to generators, fixing the conversion.

Change-Id: I1bd8575b52c624ebff2711800b99376bb083fa6a

* misc: Updating documentation.

Change-Id: I94cd0ad13116ff085d46ab61ac0cca938790809d

* misc: Removing useless imports.

Change-Id: I0db990512ff222c46c38b1482677436481909978