misc: Fixing the interface for icache and classic caches.
Fixing the issue with classic caches that break if they are not
connected to anything. The previous connect_icache worked with
ruby caches but not classic caches. This issue is fixed here and
it now works with both ruby and classic caches. I also fixed a bug
in complex_generator_core.py source code.

Change-Id: I5c59dd48b931020f8bbe728e6a971cd6b3f33b33