dev,arch-riscv: Introduced new LupV platform
This is a RISCV-based platform derived from the HiFive platform
containing the LupIO devices and will continue to be modified as more
devices are introduced. The LupIO devices are a collection of processor
agnostic and easily implemented virtual devices. Details about the
specifications of all eight LupIO devices can be found here:
https://gitlab.com/luplab/lupio/lupio-specs/-/tree/main

Information about how to build a RISCV full system with the LupIO-RTC
can be found here:
https://github.com/darchr/lupio-gem5/tree/configs

Change-Id: Ie15a217c9ab4e33d112acc3f498ba5b2eb4bb7d5