arch-riscv,dev: Make DTB not depend on python name
Currently, the RISC-V device tree generation assumes that you have very
specific variable names for the instantiated SimObjects in your config
file. For instance, it assumes that `system.cpu` is a list of BaseCPUs.
This change instead uses the state that's passed around with the device
tree generation to figure out how many CPUs there should be and uses a
unique string as the key instead of the SimObject id.

Change-Id: I4a7c670428e7068862f991577787644a8f91c72c
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>