configs: Update DTB generation for RISC-V board
This change moves the DTB generation into the board instead of using the
implicit DTB generation in many of the SimObjects. The problem with
using the implicit DTB generation is that it assumes things are named a
specific way (e.g., system.cpu). Since the component library doesn't
work that way, we need to reimplement the DTB generation. This has the
added benefit that the DTB generation can be specialized for RISC-V.

Change-Id: Ia018a812dd44a977bd41ceb5b9fed0228ff3614c
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>