scons: emit python stubfiles
Stubfiles in python allow you or your IDE to track the types of python
objects even without the full source code. This is useful for many of
gem5's objects since the source code is embedded in the gem5 binary
where IDEs can't easily see it.

This changeset adds a way to generate the stubfiles as gem5 is being
built. The files are generated to a directory under build/ and after they
are generated, you can point your IDE to that path. This requires the
stubgen program (from mypy) to be installed and does nothing if it's
not available.

Change-Id: I73966c8621d0a7a14060758d322c45df57b71102
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>