configs: Replace buildEnv with getRubyProtocol
Instead of using `buildEnv['PROTOCOL']` use the new function
`getRubyProtocol()`.

*User-facing change!*
We are deprecating the use of buildEnv['PROTOCOL'] when transitioning to
enable multiple Ruby protocols in a single binary. If you have any
configuration scripts that rely on buildEnv['PROTOCOL'] they should be
updated as shown below.

`buildEnv['PROTOCOL']` -> `getRubyProtocol()`

Change-Id: Ic4d010f93a027be03767d5306c2eee6c13ed9511
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>