python: Add command line option for ruby protocol
This command line option for the `main_ruby_protocol` will allow scripts
which assume a single protocol is compiled to work without any other
changes even if more protocols are included in the binary.

If the parameter isn't included and the binary only has one ruby
protocol then we'll default to that protocol. A warning is raised if
there is no main protocol specfied, and it will default to the "first"
compiled protocol.

This follows the implementation for multiple ISAs [1].

[1] https://gem5-review.googlesource.com/c/public/gem5/+/55423

Change-Id: I9a67345274e6b61d72c1bd4db47e3ac3b4e7d020
Signed-off-by: Jason Lowe-Power <jason@lowepower.com>