arch-riscv: Always use undisturbed tail policy
- Tail policy = 0 means tail elements are undisturbed,
  Tail policy = 1 means tail elements could undeterministically
be either undisturbed or all 1s.
  The ISA promised that the tail elements do not matter when
tail policy = 1.
  However, in the binary, if there were no decoding errors, the
tail policy = 1 and an instruction requires the tail elements
to be undisturbed.

- To implement undisturbed tail policy, the old content of Vd
must be copied to this inst Vd. This is required due to vector
reg renaming. Newly allocated vec reg is all zeros. The Vd of
the inst could be renamed to a new physical register, making
the physical register of Vd and the physical register of Vd
before renaming different. Hence the need for copying the
elements from old to the new register(s).

- Using Vd as a source register should not be expensive as
I'd expect a vector loop to process the tail elements at most in
one iteration per loop. Essentially, if a microop knows that it
will process at least one tail element, it will set Vd as a source
register as well as the dest register.

Signed-off-by: Hoa Nguyen <hoanguyen@ucdavis.edu>