stdlib,arch-riscv: Restore certain stdlib and arch files
The original implementation of the U74CPU required creating a new
simobject in riscv, and a new CPUType in the stdlib. Since a proper
workaround has been implemented, the files affected by the original
implementation can be restored to their original state. These files
are restored to commit 23b2561bf668d42b51b36ed2084a499ab46996b3, which
was the head of the develop branch of darchr/gem5 at the time this
branch was created.

Change-Id: I55fec97a6ff85c79fc4a4e166a0a07f51c012a65