stdlib: Use 4 cores in FS mode for HiFiveUnmatched board
If the HiFiveBoard object is created in FS mode, 4 cores will be
generated by _create_cores(). FS mode with 4 cores has been tested to
work. It took 2.5 hours real time to fully boot and reach the login
screen on Amarillo. The nproc command and /proc/cpuinfo confirmed that
there are 4 cores. The cpuinfo was nearly identical to the hardware
HiFiveUnmatched, with the  exception being SV39 MMU for hardware
and SV48 MMU for the gem5 board.

Change-Id: I7d4bd74cfbd5c12043ebb7feaebba34fc873901b