stdlib: Pass cache parameters to caches in HiFive Unmatched
Pass the cache associativity parameters to L1I, L1D, and L2 caches.
Prior, these parameters were set in the AbstractTwoLevelCacheHierarchy
parent class but were not passed to the __init__ functions for the
caches. The associativities are now manually set to 4, 8, and 16 for
L1I, L1D, and L2 caches respectively, as defined by the official
manual for the board. Prior, the defaults were used, which were 8, 8,
and 16 respectively at this time.

Change-Id: Ic5ddca9fe4c938bfac83cc2a09baa0ff4898121c