util,stdlib: fixed conflict exit events after all targets encountered
In LoopPointManager, the original appoarch raised conflict exit events
when all targets were encountered. Now it raises SIMPOINT_BEGIN with
exitSimLoopNow() and raises EXIT with exitSimLoop to fix this issue.

In stdlib LoopPoints, it now has a temporary testing function.
It doesn't take in anything when creating a LoopPoints module object,
instead, it takes in files depend on taking checkpoints or restoring
checkpoints.