stdlib: Added stdlib LoopPoint classes
The stdlib LoopPoint classes have to work with the LoopPoint probes and
the LoopPointManger. They are all added for running the LoopPoint
sampling method.
The stdlib LoopPoint has three classes, an abstract class and two
subclasses.
The LoopPointRestore is responsible for profiling restoring related
points.
The LoopPointCheckpoint is responsible for profiling checkpoints.
The BaseLoopPoint stores the general LoopPoints information and setup
the LoopPointManager and LoopPoint probes to each core.

Change-Id: Ie10d11e6b33189245997d3ab8e9f75cf03d76a39