base, cpu: add statExecuted and committed InstType to base
statExecutedInstType of SimpleCPU and committedInstType of MinorCPU
track the same statistics. statExecutedInstType is a 1D vector tied
to an execution context thread. committedInstType is a 2D vector, with
the first dimension referring to the thread ID. These two stat vectors
have been unified into BaseCPU::CommitCPUStats::committedInstType as
a 1D vector that is tied to a thread ID via the committedStats vector.

Change-Id: I828fec7b896adc7896dde3c64bcd2d285f0e3dcb