misc: PB test board - ran tests, pushing output files
The following benchmarks were run with simple-config.py:
riscv-print-this, riscv-bubblesort, riscv-floatmm, riscv-hello

Each benchmark was run with different rows of a size 16 foldover PB design, for a total of 32 unique board configurations. For riscv-print-this,
riscv-bubblesort, and run 0_0_16 of riscv-floatmm, each configuration had 5 runs. For the rest of the riscv-floatmm configurations and all of
the riscv-hello configurations, each configuration had 3 runs.

This commit pushes the JSON files for each configuration, the output files of each run, and some minor changes to the files used to run tests.

Change-Id: I69e0be6df2312a910fa125ec86061156fab9e678