misc: Added Yates algorithm files to analyze full factorial run
The Yates algorithm is a method to generate estimates for the effect of each parameter and interactions between parameters. To use it, we have to
access the output stats files in a order different from the order they were generated in. Specifically, the binary values of the run numbers are
flipped. This commit adds a script which generates a JSON that maps between the original and flipped number (in base 10), as well as a generated
JSON for number of parameters = 13.

Change-Id: I5f10df6790a4c2a42c37d40dbd89a77d5ee4d300