ext: outgoing bridge supports limited func accesses
This change adds a lot of new features to the gem5-SST bridge.
The following list of changes describes each of them:
1. The bridge now supports limited functional accesses, which
enables users to work with vda devices in ARM/RISCV systems.
2. There are new methods added to the bridge which allows future
extension of the bridge to swap ports to enable delayed INIT
phase for SST. This will allow the bridge to start the system in
gem5 and then swap the memory ports to SST at RUN phase.
3. A lot of stray debug statements/cout statements were removed
in this patch.
4. New documentation was added to the SST component for gem5.

Signed-off-by: Kaustav Goswami <kggoswami@ucdavis.edu>